Extract shared config model
