+++ crates/buttond/Cargo.toml
@@ -17,6 +17,7 @@ serde = { version = "1.0.217", features = ["derive"] }
 serde_yaml = "0.9.34"
 tracing = "0.1.41"
 tracing-subscriber = { version = "0.3.20", features = ["fmt", "env-filter"] }
+config-model = { path = "../config-model" }
 
 [dev-dependencies]
 tempfile = "3.13.0"
