+++ crates/wifi-manager/src/web.rs
@@ -159,8 +159,10 @@ async fn monitor_connection(state: UiState, ssid: String) {
                     return;
                 }
                 Ok(false) => {}
-                Err(err) => warn!(error = ?err, "gateway reachability check failed while monitoring"),
-            }
+                Err(err) => {
+                    warn!(error = ?err, "gateway reachability check failed while monitoring")
+                }
+            },
             Ok(false) => {}
             Err(err) => warn!(error = ?err, "device connectivity check failed while monitoring"),
         }
